refactor: structured and self-contained prepare (#531)
