+++ lib/prepare/loadConfig.js
