+++ lib/util/index.js
@@ -1,11 +1,6 @@
-const spawn = require('cross-spawn')
 const parseHeaders = require('./parseHeaders')
 
-exports.encodePath = path => {
-  return path.split('/').map(item => encodeURIComponent(item)).join('/')
-}
-
-exports.normalizeHeadTag = tag => {
+exports.normalizeHeadTag = function (tag) {
   if (typeof tag === 'string') {
     tag = [tag]
   }
@@ -18,7 +13,7 @@ exports.normalizeHeadTag = tag => {
   }
 }
 
-exports.applyUserWebpackConfig = (userConfig, config, isServer) => {
+exports.applyUserWebpackConfig = function (userConfig, config, isServer) {
   const merge = require('webpack-merge')
   if (typeof userConfig === 'object') {
     return merge(config, userConfig)
@@ -32,7 +27,7 @@ exports.applyUserWebpackConfig = (userConfig, config, isServer) => {
   return config
 }
 
-exports.inferTitle = frontmatter => {
+exports.inferTitle = function (frontmatter) {
   if (frontmatter.data.home) {
     return 'Home'
   }
@@ -45,7 +40,7 @@ exports.inferTitle = frontmatter => {
   }
 }
 
-exports.parseFrontmatter = content => {
+exports.parseFrontmatter = function (content) {
   const matter = require('gray-matter')
   const toml = require('toml')
 
@@ -61,7 +56,7 @@ exports.parseFrontmatter = content => {
 const LRU = require('lru-cache')
 const cache = LRU({ max: 1000 })
 
-exports.extractHeaders = (content, include = [], md) => {
+exports.extractHeaders = function (content, include = [], md) {
   const key = content + include.join(',')
   const hit = cache.get(key)
   if (hit) {
@@ -85,8 +80,4 @@ exports.extractHeaders = (content, include = [], md) => {
 
   cache.set(key, res)
   return res
-}
-
-exports.getGitLastUpdatedTimeStamp = filepath => {
-  return parseInt(spawn.sync('git', ['log', '-1', '--format=%ct', filepath]).stdout.toString('utf-8')) * 1000
 }
