+++ test/prepare/fixtures/docs-custom-theme/.vuepress/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  title: 'Hello VuePress',
+  description: 'Just playing around'
+}
