+++ test/prepare/fixtures/docs-custom-theme/README.md
@@ -1,0 +1,1 @@
+# Simple Docs
