+++ test/prepare/fixtures/docs-simple-config/.vuepress/config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  title: 'Hello VuePress',
+  description: 'Just playing around',
+  dest: 'vuepress',
+  base: 'vuepress',
+  head: [
+    ['link', { rel: 'icon', href: '/logo.png' }]
+  ]
+}
