+++ test/prepare/fixtures/docs-simple-config/README.md
@@ -1,0 +1,1 @@
+# Simple Docs
