+++ test/prepare/fixtures/docs-simple/README.md
@@ -1,0 +1,1 @@
+# Simple Docs
